Switch to using JCommander for the params/arguments system.
This system is very complex and we can make it as complex or as simple
as we want. This will save us time when we want to parse commands and
since it takes a while to write the code to parse the commands, this way
provides a lot more options. I'm excited for this.