First stab at tab complete of jail commands, #77
Right now, the first option of the jail command completes and checks for
proper permission and all that good stuff. Good way to get started with
commands but that's all that has been completed right now.

Plan is that the command interface will get a tabComplete method which
all the classes that implement it will handle. Then if a tabComplete
happens on that command we'll let the command handle it, so the command
handlers don't get cluttered.